Initial commit
Includes some structure for the oremda python package, and a Dockerfile
for a base oremda image, along with an example. The example works in
Docker and in Singularity, although there are some errors when Singularity
runs (even though it appears to perform the operations on the memory like
it is supposed to).

A unified interface for running singularity and docker containers via
a python class is in progress.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>