Add ThreadPoolSingleton to reuse and limit threads
If we launch threads in our program using this singleton, it will
re-use threads in the thread pool (so we won't have as much overhead
of destroying and creating them), and it will limit the number of
threads that the program is using to the cpu count.

This will limit the number of threads.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>