Add a decorator to simplify creating an operator
We can now just decorate a function, and it will do the work
of subclassing and setting the kernel function under the hood.

The name of the function is important, because the operator strings
in the pipeline will need to match.

The decorator actually returns an object of the subclass, rather
than the function.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>