Stop operators cleanly
This gets the pipeline to send a message to the operators indicating
that they should terminate. The operators then exit cleanly.

This removes the need to kill the docker containers, which is a
little trickier for other containers, such as singularity, since
we are using the singularity-cli which doesn't give us the singularity
subprocess pid.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>