Add utility for starting plasma store
This context manager is loosely based upon [plasma's `start_plasma_store()`](https://github.com/apache/arrow/blob/15b7ae1521a5010964aa475e7600dda80c03bc89/python/pyarrow/plasma.py#L82),
but unlike plasma's `start_plasma_store()`, this one allows us to specify
the socket path. It also removes a lot of parts we don't need.

This is used to close the plasma store cleanly. And it is now being done as
part of the loader, which is functioning more like a driver at this point...

Closing the plasma store cleanly is important for singularity, where it is
harder to kill the processes.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>