Run singularity client without instances
This updates the singularity client to run without instances, as
they are more complicated and will not be supported in other container
technologies.

This means, however, that we will have a harder time killing singularity
processes, as the singularity python API does not provide any way to
kill the process, nor does it provide things like the subprocess id.

However, the containers have been updated so that they close cleanly now.
So we do not need to worry as much about that for singularity.

This also updates runner.py so that it will work with singularity also.
It seems to be working.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>