Add images and labels to container classes
This separates out classes into different files for the container
interfaces. It also adds images, which are used to get labels.

There is additionally an "oremda_labels" property that returns a
nested dict of oremda labels.

A docker_to_singularity.py script was added to build a singularity
image from a docker image, with the docker labels added to the
singularity labels (which is currently not done automatically).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>