Add a basic stateful volume rendering example
This adds a vtk reader operator, and a stateful volume rendering
operator.

The stateful volume rendering operator is our first example of an
operator that has state. Running this operator will add a volume to
the operator's internal `vtkRenderWindow`. Currently, the operator will
just write out a PNG file of what it rendered last.

No metadata such as spacing and origin is currently being considered.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>