Have pipeline pass output queue to operators
This tells the operators which queue they should put their output
in. This will avoid confusion in the case of multiple pipelines
running, as the pipeline id is in the queue name (so it will be
passed back to the correct pipeline). Also, the id of the operator
node in the pipeline is in the name as well, so that will be unique
also.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>