Change DEFAULT_OREMDA_VAR_DIR to /tmp
/run/oremda has been difficult to use because of permissions errors.
Every time a computer is restarted, the directory is deleted, and the
user has to remake the directory with sudo and change the ownership.

On top of that, singularity containers have been having issues with
this directory as well due to permissions issues. Just change it
to /tmp to fix numerous permissions issues.

We should soon be only mounting the oremda var dir from a parent
container to children containers soon, so we do not have to worry
about any security issues (the host will no longer mount the oremda
var dir into the parent container).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>