Make compatibility changes for singularity
This adds some compatibility changes so that singularity can work
again.

Namely, doing things like "self_container()", "container()", and
"mounts" is not easy in singularity, and may be impossible in shifter.
These methods were kept in Docker, but removed from the base class
so that other containers do not have to implement them.

This also fixes an issue where you must delete the singularity
bind mounts environment variable to avoid repeating the same bind mounts
in a child container.

This also adds a working directory for singularity as well, and an example
json script. Maybe in the near future we can allow docker-style image
names and get singularity to automatically find the images that match.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>