Fix up docker example
Singularity and docker are differing in that in Singularity, we are
launching child containers from a parent container, but in docker,
we are launching sibling containers from a parent container.

This means the mounts will be different. Treat it as such.

This was tested with both Docker and Singularity and appeared to work.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>