Add MPI Messenger and singularity example
The MPI messenger allows messages to be sent and received. It is
currently just using pickling, because that is easy, but the efficiency
can certainly be improved in the future.

OREMDA's messaging with MPI is designed in a somewhat complicated way
that makes creating operators easy. The outer runner is started with
MPI. On the node with rank 0, the runner runs the pipeline, which sends
out instructions via MPI to remote operators. On nodes with rank != 0,
the outer runner starts its child operator container, and the outer
runner receives MPI messages from the pipeline and forwards them to
the child container via message queues and plasma. It then receives
the output from the child container and sends them back via MPI to the
pipeline. This is done so that the child operator does not need to be
built with MPI, but the same operator container used locally can also
be used on HPC. The communication with message queues and plasma can
also avoid any deep copies of the data if done correctly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>