Add ability to resolve singularity paths
This will first try the path given to it (in case it is a full path
to a singularity image), and if the path exists, it will use it.

If the path does not exist, it assumes that it may be a docker image,
and it uses a function to convert it to a singularity image name.
For example: "oremda/runner:latest" => "oremda_runner_latest.sif".

The user can also set an images directory on the singularity client
which tells singularity where to look for the images.

This simplifies running singularity so that we do not need to specify
the full path to the images in the pipeline.json file, and we can instead
use the same docker names, and just supply singularity with the directory
to search.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>