Add timeout fix to test()
For some reason (which would be really nice to know and understand),
we need to test() the mpi request multiple times in a timeout loop. If
we don't, and we have more ranks than say 5, we start getting some
requests where the MPI message has been received, but test() fails
anyways. I have no idea why, maybe this is related to this issue:
https://stackoverflow.com/questions/26870796/mpi-test-isnt-picking-up-a-completed-mpi-send

But I don't fully understand what he is saying.

Anyways, adding multiple tests and a timeout seems to fix the issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>